requirements.txt
absl-py==0.9.0
astor==0.8.1
click==7.1.1
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.8
gast==0.2.2
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
imageio==2.8.0
imgaug==0.4.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.2
opencv-python==4.2.0.34
opt-einsum==3.2.0
Pillow==7.1.1
protobuf==3.11.3
pybase64==1.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
Werkzeug==1.0.1
wrapt==1.12.1
